+++ android/src/main/java/com/baidu/bmfmap/map/maphandler/CustomMapHandler.java
@@ -15,9 +15,10 @@ import com.baidu.mapapi.map.MapCustomStyleOptions;
 import android.content.Context;
 import android.text.TextUtils;
 import android.util.Log;
+
+import io.flutter.FlutterInjector;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.view.FlutterMain;
 
 public class CustomMapHandler extends BMapHandler {
     
@@ -118,7 +119,7 @@ public class CustomMapHandler extends BMapHandler {
         String parentPath = null;
         String customStyleFileName = null;
         try {
-            customStyleFileName = FlutterMain.getLookupKeyForAsset(customStyleFilePath);
+            customStyleFileName = FlutterInjector.instance().flutterLoader().getLookupKeyForAsset(customStyleFilePath);
             inputStream = context.getAssets().open(customStyleFileName);
             byte[] buffer = new byte[inputStream.available()];
             inputStream.read(buffer);
@@ -186,10 +187,12 @@ public class CustomMapHandler extends BMapHandler {
         if (mMapController != null && mMapController.getFlutterMapViewWrapper() != null) {
             mMapController.getFlutterMapViewWrapper().setMapCustomStyle(mapCustomStyleOptions,
                     new CustomMapStyleCallBack() {
+                        boolean isPreloaded = false;
+                        String preloadPath = null;
                 @Override
                 public boolean onPreLoadLastCustomMapStyle(String path) {
-                    reslutMap.put("preloadPath", path);
-                    result.success(reslutMap);
+                    isPreloaded = true;
+                    preloadPath = path;
                     return false;
                 }
 
@@ -197,7 +200,14 @@ public class CustomMapHandler extends BMapHandler {
                 public boolean onCustomMapStyleLoadSuccess(boolean b, String path) {
                     reslutMap.put("styHasUpdate", String.valueOf(b));
                     reslutMap.put("successPath", path);
+                    if (isPreloaded) {
+                        reslutMap.put("preloadPath", path);
+                    }
                     result.success(reslutMap);
+                    // 防止在线加载后Flutter不刷新地图
+                    if (mMapController.getBaiduMap() != null) {
+                        mMapController.getBaiduMap().mapRefresh();
+                    }
                     return false;
                 }
 
@@ -206,6 +216,9 @@ public class CustomMapHandler extends BMapHandler {
                     String sStatus = String.valueOf(status);
                     reslutMap.put("errorCode", sStatus);
                     reslutMap.put("errorPath", path);
+                    if (isPreloaded) {
+                        reslutMap.put("preloadPath", path);
+                    }
                     result.success(reslutMap);
                     return false;
                 }